+++ App/controllers/recommendation.py
@@ -2,27 +2,6 @@ from App.models import Recommendation, Student
 from App.database import db
 from sqlalchemy.exc import IntegrityError
 
-def create_recommendation(sentFromStaffID, sentToStudentID, recURL):
-    newrec = Recommendation(sentFromStaffID=sentFromStaffID, sentToStudentID=sentToStudentID, recURL=recURL)
-    return newrec
-
-def send_recommendation(sentFromStaffID, sentToStudentID, recURL):
-    student = Student.query.get(sentToStudentID)
-    newrec = create_recommendation(sentFromStaffID, sentToStudentID, recURL)
-    try:
-        db.session.add(newrec)
-        db.session.commit()
-    except IntegrityError:
-        db.session.rollback()
-        return None
-    student.recommendationList.append(newrec)
-    try:
-        db.session.add(student)
-        db.session.commit()
-    except IntegrityError:
-        db.session.rollback()
-        return None
-    return student
 
 def get_all_recommendations():
     return Recommendation.query.all()
